src/discord/requirements.txt
python-dotenv
faiss-cpu
sentence-transformers
discord
openai
tiktoken
google-generativeai
